Provide a means of checking based on the hour
Orcfax currently assumes a publication on the hour. If a publication
has not been produced based on the previous hour that ticked by then
it should be pushed on chain. The current heauristic takes into
account the previous publication as well as the feed's interval.
Tests have been provided to try and capture as many cases as
possible.

May require some refactoring to improve the loops involved with the
time-deltas. This should be a pretty decent first-cut though (
providing it works as anticipated).