Ensure slot isn't updated by price monitoring
Price monitoring is a long-running process and needs to query the
chain more often. If the slot hasn't been updated we can exit early.
If not, we can query the chain.

We may reconsider the use of slot number in the future. It's impact is
only small and we can probably query Kupo as much as we want.