Make sure ValueErrors are handled
During the Cardano Node Fork we received something other than an int
value for the slot. This caused an error we didn't anticipate and it caused
the price monitor to stop running. This prevents that and returns a known
error to the caller. Tenacity ensures that code continues to run.